backend/requirements.txt
livekit-agents[assemblyai,silero,turn-detector]==1.2.17
livekit-plugins-speechmatics==1.2.17
livekit-plugins-noise-cancellation==0.2.5
python-dotenv==1.2.1
cryptography==46.0.3
langchain-openai>=0.1.0
fastapi==0.115.0
uvicorn[standard]==0.32.0
websockets==14.1
python-multipart==0.0.12
huggingface-hub>=0.34.0,<1.0
numpy<2.0
protobuf<6.0,>=3.20.3
watchdog>=3.0.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
redis>=5.0.0
sentence-transformers>=2.2.0

